Fix chunksize not loading for peeler when launched from gui
In `MainWindow.run_peeler`, `peeler.change_params` is called without passing in the `chunksize`. This sets the Peeler's `chunksize` to the default value of 1024.

To pass the correct `chunksize` to `peeler.change_params`, the `chunksize` must be retrieved. If the `chunksize` was set in `dialog_fullchain_params` this session, it would be possible to retrieve it from there. However, if a data set was loaded with a previously initialized catalogue, the default value found in `dialog_fullchain_params` will be incorrect.

So, to solve this problem, this commit first adds `chunksize` to `CatalogueConstructor.make_catalogue` so that it can be saved by `CatalogueConstructor.make_catalogue_for_peeler` and `DataIO.save_catalogue`. Second, the commit extracts the `chunksize` from `initial_catalogue` after it has been loaded by `DataIO.load_catalogue` inside `MainWindow.run_peeler` and passes it with the other parameters to `peeler.change_params`.

Extraction of the new `chunksize` parameter from `initial_catalogue` is wrapped in a `try` statement for backwards compatibility with old catalogues created before this fix.