Bug fix: Load default for lostfront_chunksize in set_preprocessor_params
The docstring for `CatalogueConstructor.set_preprocessor_params` claims that a reasonable value for `lostfront_chunksize` is set by default if `None` is provided. However, the code that sets this value is located in `SignalPreprocessor_base.change_params` and may not be executed before it's needed.

For example, if `estimate_signals_noise` is run immediately after `set_preprocessor_params(lostfront_chunksize = None, ...)` (as is the case with `apply_all_catalogue_steps`), a `TypeError` occurs because `lostfront_chunksize` is still `None`.

This commit duplicates the code for setting the default value into `set_preprocessor_params` to avoid this bug.